+++ lib/mir-fmm/src/reference_count.rs
@@ -5,8 +5,6 @@ pub mod heap;
 pub mod pointer;
 pub mod record;
 mod record_utilities;
-mod sync;
 pub mod variant;
 
 pub use expression::*;
-pub use sync::*;
