+++ lib/mir-fmm/src/reference_count/expression.rs
@@ -76,3 +76,24 @@ pub fn drop(
 
     Ok(())
 }
+
+pub fn synchronize(
+    builder: &fmm::build::InstructionBuilder,
+    expression: &fmm::build::TypedExpression,
+    type_: &mir::types::Type,
+    _types: &FnvHashMap<String, mir::types::RecordBody>,
+) -> Result<(), CompileError> {
+    match type_ {
+        mir::types::Type::ByteString => {
+            pointer::synchronize(builder, expression)?;
+        }
+        mir::types::Type::Function(_) => {
+            function::synchronize(builder, expression)?;
+        }
+        mir::types::Type::Record(_) => todo!(),
+        mir::types::Type::Variant => todo!(),
+        mir::types::Type::Boolean | mir::types::Type::None | mir::types::Type::Number => {}
+    }
+
+    Ok(())
+}
